requirements.txt
streamlit>=1.32.0
tensorflow>=2.15.0
numpy>=1.24.0
Pillow>=10.0.0
plotly>=5.18.0
streamlit-drawable-canvas>=0.9.3